move ArtifactThemeReference class to internal.io.reference package
